libraries/python/message-history-manager/pyproject.toml
[project]
name = "message-history-manager"
version = "0.1.0"
description = "Manages message history for an OpenAI agent"
authors = [{ name = "Semantic Workbench Team" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.85,<2.0",
    "openai-client>=0.1.0",
    "pydantic>=2.10,<3.0",
    "python-dotenv>=1.0.1,<2.0",
    "python-liquid>=2.0,<3.0",
]

[tool.uv.sources]
openai-client = { path = "../openai-client", editable = true }

[dependency-groups]
dev = ["pyright>=1.1.401", "pytest>=8.4.0", "pytest-asyncio>=1.0.0"]
examples = ["pydantic-ai[openai,athropic]>=0.2,<0.3"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = ["-vv"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s | %(levelname)-7s | %(name)s | %(message)s"
testpaths = ["test"]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
